requirements.txt
imutils==0.5.4
django==3.2.9
django-bootstrap3==21.1
django-phonenumber-field==6.0.0
phonenumbers==8.12.37
dlib==19.19.0
geocoder==1.38.1
geopy==2.2.0
scipy==1.7.1
twilio==7.3.1
gunicorn==20.0.4
whitenoise==5.3.0
djangorestframework==3.12.4
django-cors-headers==3.10.0
django-cors-middleware==1.5.0
opencv-python-headless